split up request factory for searches
